feat: Decoupled system init from executors.
