+++ mava/systems/jax/mappo/system.py
@@ -34,6 +34,11 @@ class MAPPOSystem(System):
         default_params = MAPPODefaultConfig()
 
         # Default system processes
+        # System initialization
+        system_init = DesignSpec(
+            environment_spec=building.EnvironmentSpec, system_init=building.SystemInit
+        ).get()
+
         # Executor
         executor_process = DesignSpec(
             executor_init=executing.ExecutorInit,
@@ -71,6 +76,7 @@ class MAPPOSystem(System):
         ).get()
 
         system = DesignSpec(
+            **system_init,
             **data_server_process,
             **parameter_server_process,
             **executor_process,
